+++ src/middleware/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  json: require('./json'),
+  redirect: require('./redirect'),
+  throttling: require('./throttling')
+};
